WIP: refactor: moved the get_file_name_from_file_list_metadata get_file_rel_path_from_file_list_metadata
